+++ kiclearance/__main__.py
@@ -37,5 +37,6 @@ for o, a in opts:
 # Run script
 table_file = look_for_clearance_table_file(project_folder, table_name)
 project_name = look_for_kicad_project(project_folder, project_name)
+update_net_classes_from_kicad_project_to_clearance_table_file(kicad_project_name=project_name, clearance_table_file=table_file, folder=project_folder)
 table_data = parse_excel_table(table_file)
 write_design_rule_file(table_data, project_folder, project_name, factor_inner_layers=factor_inner_layers, min_track_distance=min_track_distance)