+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Introduce clearance table generation and clearance automatic table update from kicad project
+
 ### Fixed
  - reduce rules in case of factor_inner_layers is 1.0